fix: robot_is_ready_for_new_command now returns explicit floats instead of bools (that were converted implicitly to floats)

+++ rl_quadrupeds/quadrupeds_mdp/observations/state.py
@@ -9,4 +9,4 @@ def robot_is_ready_for_new_command(
     if not hasattr(env, "is_ready_for_new_command"):
         raise ValueError("The environment does not have 'is_ready_for_new_command' attribute.")
     
-    return env.is_ready_for_new_command
+    return env.is_ready_for_new_command.float()
